+++ website/src/components/Submission/FileUpload/UploadComponent.tsx
@@ -1,142 +1,8 @@
-import { useCallback, useEffect, useRef, useState, type SVGProps, type ForwardRefExoticComponent } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { toast } from 'react-toastify';
-import * as XLSX from 'xlsx';
 
+import type { FileKind, ProcessedFile } from './fileProcessing.ts';
 import useClientFlag from '../../../hooks/isClient.ts';
-import MaterialSymbolsLightDataTableOutline from '~icons/material-symbols-light/data-table-outline';
-import PhDnaLight from '~icons/ph/dna-light';
-
-type Icon = ForwardRefExoticComponent<SVGProps<SVGSVGElement>>;
-
-type FileKind = {
-    type: 'metadata' | 'fasta';
-    icon: Icon;
-    supportedExtensions: string[];
-    processRawFile: (file: File) => Promise<ProcessedFile | Error>;
-};
-
-export const METADATA_FILE_KIND: FileKind = {
-    type: 'metadata',
-    icon: MaterialSymbolsLightDataTableOutline,
-    supportedExtensions: ['tsv', 'xlsx', 'xls'],
-    processRawFile: async (file: File) => {
-        switch (file.type) {
-            case 'application/vnd.ms-excel':
-            case 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': {
-                const f = new ExcelFile(file);
-                try {
-                    await f.init();
-                } catch (err) {
-                    return err as Error;
-                }
-                return f;
-            }
-            default:
-                return new RawFile(file);
-        }
-    },
-};
-
-export const FASTA_FILE_KIND: FileKind = {
-    type: 'fasta',
-    icon: PhDnaLight,
-    supportedExtensions: ['fasta'],
-    processRawFile: (file) => Promise.resolve(new RawFile(file)),
-};
-
-interface ProcessedFile {
-    /* The file containing the data (might be processed, only exists in memory) */
-    inner(): File;
-
-    /* The handle to the file on disk. */
-    handle(): File;
-
-    /* Warnings that came up during file processing. */
-    warnings(): string[];
-}
-
-class RawFile implements ProcessedFile {
-    private innerFile: File;
-
-    constructor(file: File) {
-        this.innerFile = file;
-    }
-
-    inner(): File {
-        return this.innerFile;
-    }
-    handle(): File {
-        return this.innerFile;
-    }
-
-    warnings(): string[] {
-        return [];
-    }
-}
-
-class ExcelFile implements ProcessedFile {
-    private originalFile: File;
-    private tsvFile: File | undefined;
-    private processingWarnings: string[];
-
-    constructor(excelFile: File) {
-        // assumes that the given file is actually an execel file.
-        this.originalFile = excelFile;
-        this.processingWarnings = [];
-    }
-
-    async init() {
-        const arrayBuffer = await this.originalFile.arrayBuffer();
-        const workbook = XLSX.read(arrayBuffer, {
-            type: 'array',
-            cellDates: true, // parse date cells into actual date objects (as opposed to numbers)
-            dateNF: 'yyyy-mm-dd', // use this format to 'render' date cells
-        });
-
-        const firstSheetName = workbook.SheetNames[0];
-        const sheet = workbook.Sheets[firstSheetName];
-        const tsvContent = XLSX.utils.sheet_to_csv(sheet, {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            FS: '\t',
-            blankrows: false,
-        });
-        const rowCount = tsvContent.split('\n').length - 1;
-        if (rowCount <= 0) {
-            throw new Error(`Sheet ${firstSheetName} is empty.`);
-        }
-        /* eslint-disable no-console */
-        console.log('SHEET NAMES:');
-        console.log(JSON.stringify(workbook.SheetNames));
-        console.log('-----------------------------------');
-        console.log(tsvContent);
-        console.log('-----------------------------------');
-        /* eslint-enable no-console */
-
-        const tsvBlob = new Blob([tsvContent], { type: 'text/tab-separated-values' });
-        const tsvFile = new File([tsvBlob], 'converted.tsv', { type: 'text/tab-separated-values' });
-        this.tsvFile = tsvFile;
-        if (workbook.SheetNames.length > 1) {
-            this.processingWarnings.push(
-                `The file contains ${workbook.SheetNames.length} sheets, only the first sheet (${firstSheetName}; ${rowCount} rows) was processed.`,
-            );
-        }
-    }
-
-    inner(): File {
-        if (this.tsvFile === undefined) {
-            throw new Error('file was not initialized');
-        }
-        return this.tsvFile;
-    }
-
-    handle(): File {
-        return this.originalFile;
-    }
-
-    warnings(): string[] {
-        return this.processingWarnings;
-    }
-}
 
 export const UploadComponent = ({
     setFile,
@@ -161,7 +27,7 @@ export const UploadComponent = ({
                 processingResult = null;
             }
             if (processingResult?.warnings().length) {
-                toast.warn(processingResult.warnings().join(" "));
+                toast.warn(processingResult.warnings().join(' '));
             }
             setFile(processingResult ? processingResult.inner() : null);
             rawSetMyFile(processingResult);
